feat: enable custom javascript
To use this feature, add the `custom-javascript`-option to the
configuration of the Microblog-Action and set it to a path relative to
the repo root.

The contents of the file will be inlined at the bottom of each HTML
page.

Example:

```yaml
- name: Build Microblog
  uses: herschel666/microblog-action@v0.4
  with:
    custom-javascript: 'assets/scripts.js'
```

Keep in mind, that there's no transpiling or minifying in place. The
Javascript will be inserted as-is.